server/requirements.txt
# Minimal AstroNova Backend Dependencies
#
# Quick start:
#   pip install -r requirements.txt
#   python app.py

flask>=2.3.0,<4.0.0
flask-cors>=5.0.0,<7.0.0
flask-limiter>=3.5.0,<4.0.0
Flask-Babel==4.0.0

# Apple Sign-In token validation
pyjwt>=2.8.0,<3.0.0
cryptography>=41.0.0,<43.0.0
requests>=2.31.0,<3.0.0

# AI chat services (Gemini preferred, OpenAI fallback)
google-generativeai>=0.8.0,<1.0.0
openai>=1.0.0,<2.0.0

# Required: Swiss Ephemeris (high-accuracy planetary positions)
pyswisseph>=2.10.0,<3.0.0

# Optional: geocoding for location search (fallback provided if missing)
geopy>=2.4.0,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-benchmark>=4.0.0,<5.0.0
pytest-cov>=4.1.0,<5.0.0
memory-profiler>=0.61.0,<1.0.0
freezegun>=1.5.0,<2.0.0
